requirements.txt
amocrm-api==2.6.1
asgiref==3.8.1
blinker==1.9.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
Django==5.1.2
djangorestframework==3.15.2
Flask==3.1.0
Flask-Cors==5.0.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
psycopg==3.2.3
psycopg-binary==3.2.3
psycopg2==2.9.10
psycopg2-binary==2.9.10
PyJWT==2.10.0
python-dotenv==1.0.1
python-slugify==8.0.4
requests==2.32.3
sqlparse==0.5.1
text-unidecode==1.3
tzdata==2024.2
urllib3==2.2.3
Werkzeug==3.1.3
